ref: restructure common mod
Also fills out the test node impl and adds a helper to enable tracing.
